product class created
