--- Full_Stack_Demo/src/main/java/com/example/demo/repository/EmployeeRepository.java
@@ -1,9 +1,0 @@
-package com.example.demo.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.example.demo.entity.Employee;
-
-public interface EmployeeRepository extends JpaRepository<Employee, Long>{
-
-}
