+++ src/main/java/com/example/demo/entity/Product.java
@@ -1,0 +1,5 @@
+package com.example.demo.entity;
+
+public class Product {
+
+}
